requirements.txt
pillow
pandas
matplotlib
seaborn
plotly
pywebview